Refactor defined attribute rewriting; move statement rewriter stuff into its own package in core.
