+++ core/src/main/java/org/jdbi/v3/PreparedBatch.java
@@ -25,8 +25,8 @@ import java.util.function.Function;
 
 import org.jdbi.v3.exceptions.UnableToCreateStatementException;
 import org.jdbi.v3.exceptions.UnableToExecuteStatementException;
+import org.jdbi.v3.rewriter.RewrittenStatement;
 import org.jdbi.v3.tweak.ColumnMapper;
-import org.jdbi.v3.tweak.RewrittenStatement;
 import org.jdbi.v3.tweak.RowMapper;
 import org.jdbi.v3.tweak.StatementBuilder;
 import org.jdbi.v3.tweak.StatementCustomizer;
