+++ core/src/main/java/org/jdbi/v3/SqlStatement.java
@@ -29,15 +29,15 @@ import java.util.Map;
 
 import org.jdbi.v3.exceptions.UnableToCreateStatementException;
 import org.jdbi.v3.exceptions.UnableToExecuteStatementException;
+import org.jdbi.v3.rewriter.RewrittenStatement;
+import org.jdbi.v3.rewriter.StatementRewriter;
 import org.jdbi.v3.tweak.Argument;
 import org.jdbi.v3.tweak.ArgumentFactory;
 import org.jdbi.v3.tweak.CollectorFactory;
 import org.jdbi.v3.tweak.NamedArgumentFinder;
-import org.jdbi.v3.tweak.RewrittenStatement;
 import org.jdbi.v3.tweak.StatementBuilder;
 import org.jdbi.v3.tweak.StatementCustomizer;
 import org.jdbi.v3.tweak.StatementLocator;
-import org.jdbi.v3.tweak.StatementRewriter;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
