+++ core/src/main/java/org/jdbi/v3/rewriter/ParsedStatement.java
@@ -1,0 +1,36 @@
+/*
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.jdbi.v3.rewriter;
+
+import java.util.ArrayList;
+import java.util.List;
+
+class ParsedStatement {
+    String sql;
+    boolean positionalOnly = true;
+    List<String> params = new ArrayList<>();
+
+    void addNamedParamAt(String name) {
+        positionalOnly = false;
+        params.add(name);
+    }
+
+    void addPositionalParamAt() {
+        params.add("*");
+    }
+
+    String getParsedSql() {
+        return sql;
+    }
+}
