+++ core/src/main/java/org/jdbi/v3/rewriter/RewrittenStatement.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jdbi.v3.tweak;
+package org.jdbi.v3.rewriter;
 
 import java.sql.PreparedStatement;
 import java.sql.SQLException;
