+++ core/src/test/java/org/jdbi/v3/rewriter/TestEscapedCharacters.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jdbi.v3;
+package org.jdbi.v3.rewriter;
 
 import org.junit.Assert;
 import org.junit.Test;
