+++ sqlobject/src/main/java/org/jdbi/v3/sqlobject/customizers/OverrideStatementRewriterWith.java
@@ -21,10 +21,10 @@ import java.lang.annotation.Target;
 import java.lang.reflect.Constructor;
 import java.lang.reflect.Method;
 
+import org.jdbi.v3.rewriter.StatementRewriter;
 import org.jdbi.v3.sqlobject.SqlStatementCustomizer;
 import org.jdbi.v3.sqlobject.SqlStatementCustomizerFactory;
 import org.jdbi.v3.sqlobject.SqlStatementCustomizingAnnotation;
-import org.jdbi.v3.tweak.StatementRewriter;
 
 /**
  * Use this to override the statement rewriter on a sql object, May be specified on either the interface
